Remove private RawSerial object
The serial port is anyway shared between boxes. The practice of creating
box-private objects is already well described by InterruptIn, so there
is no need to have this one as well.

In addition, sharing the serial port was creating race conditions on the
UART IRQs between boxes on the EFM32GG_STK3700 target.